finish part 4

+++ part4/bloglist/controllers/blogs.js
@@ -1,6 +1,18 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+/*
+const getTokenForm = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+
+  return null
+}
+*/
 
 blogsRouter.get('/', async (request, response) => {
   /*
@@ -27,17 +39,24 @@ blogsRouter.get('/:id', async (request, response, next) => {
 
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
-  const user = await User.findById(body.userId)
-
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes,
-    user: user._id
-  })
-
+  //const token = getTokenForm(request)
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRECT)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid'})
+    }
+    
+    const user = request.user
+
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes,
+      user: user._id
+    })
+
+  
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
@@ -49,8 +68,23 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRECT)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid'})
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'You are not the authorized to delete this blog'})
+  }
+
+  
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
